Abstract enemy and patterns
