--- Assets/Scripts/Lootables/AttackMode.cs
@@ -1,1 +1,0 @@
-﻿public enum AttackMode { CritFail, Fail, Normal, CritSuccess };
