+++ Assets/Scripts/Systems/BattleHit.cs
@@ -2,6 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using DeCrawl.Utils;
+using DeCrawl.Primitives;
 
 public class BattleHit : MonoBehaviour
 {
