+++ Assets/deCrawl/Scrips/Enemies.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ff79587ae675b7c48b4d61af8bfb3887
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
