+++ Assets/deCrawl/Scrips/Enemies/EnemyBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2b34b484f16218f42b76d2f937bce2b9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
