+++ Assets/deCrawl/Scrips/Enemies/PatternEnemy.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f02fe97c2f6148248935a0c381841f47
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
