+++ Assets/deCrawl/Scrips/Enemies/PatternEnemy/AbstractEnemyPattern.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 754a5ae9781303644b832f303afafec3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
