+++ Assets/deCrawl/Scrips/Primitives/AttackMode.cs
@@ -1,0 +1,4 @@
+﻿namespace DeCrawl.Primitives
+{
+    public enum AttackMode { CritFail, Fail, Normal, CritSuccess };
+}
