+++ Assets/deCrawl/Scrips/Primitives/IAttackStats.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace DeCrawl.Primitives
+{
+    public interface IAttackStats
+    {
+        public AttackMode GetAttack(out int amount);
+    }
+}
