+++ Assets/deCrawl/Scrips/Primitives/IAttackStats.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 34b1d11feca1d974499f26c1ea5654d7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
